feat: add `QuizDemo`
