+++ resources/js/Pages/Admin/QuizDemo.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import Quiz from '@/Pages/User/Quiz.vue'
+import QuizLayout from '@/Layouts/QuizLayout.vue'
 import { converter } from '@/Helpers/Converter'
 
 defineProps<{ quiz: Quiz }>()
 </script>
 
 <template>
-  <Quiz :user-quiz="converter.quiz.adminToUser(quiz)" />
+  <QuizLayout :user-quiz="converter.quiz.adminToUser(quiz)" />
 </template>
